requirements.txt
2to3==1.0
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
boto==2.49.0
boto3==1.9.111
botocore==1.12.112
bpe==1.0
bz2file==0.98
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
cycler==0.10.0
docutils==0.14
gast==0.2.2
gensim==3.4.0
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
hypothesis==4.24.3
idna==2.8
importlib-metadata==0.18
jmespath==0.9.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
git+https://github.com/kpot/keras-transformer#egg=keras-transformer
kiwisolver==1.0.1
Markdown==3.1.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==7.0.0
mypy==0.701
mypy-extensions==0.4.1
nltk==3.4
normalise==0.1.8
numpy==1.16.4
olefile==0.46
packaging==19.0
pandas==0.24.2
pbr==5.2.1
Pillow==5.4.1
pluggy==0.12.0
protobuf==3.8.0
py==1.8.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.6.8
pytest==4.6.3
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1.1
requests==2.21.0
roman==3.2
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.3.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
stevedore==1.30.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
typed-ast==1.3.5
urllib3==1.24.1
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.1
